Add setup_test.go for modularity
